refactor(sim): PadBus - the pad drive channel as one piece PinManager and engine bridges share
An overlay bridge that owns an in-browser engine directly does not report
through PinManager; it needs the same dedup, the same derived resting level
and the same event shape to feed the same LineSensorHub. Extracted from
PinManager, which now delegates to it. No behaviour change.
